requirements.txt
# python 3.12
django>=4.2.0,<5.0
djangorestframework>=3.15.0
django-cors-headers==4.7.0
gunicorn>=20.1.0
psycopg2-binary==2.9.10
Faker==37.1.0
coverage==7.8.0
django-extensions==3.2.3
django_db_logger
drf-spectacular==0.28.0
python-decouple==3.8
requests==2.32.4
tenacity==9.1.2
faker_vehicle==0.2.0